+++ src/graphql/resolvers/brand/updateOne.ts
@@ -20,6 +20,11 @@ const updateOneBrand: T.Resolver<GraphQL.MutationUpdateOneBrandArgs, GraphQL.Bra
     data: {
       url: data.url || undefined,
       email: data.email || undefined,
+      v: data.v || undefined,
+      isPublished: typeof data.isPublished === 'boolean' ? data.isPublished : undefined,
+      added: data.added || undefined,
+      edited: data.edited || undefined,
+      published: data.published || undefined,
       avgTimeStory: data.avgTimeStory || undefined,
       avgAllTimeStory: data.avgAllTimeStory || undefined,
       updated: new Date(),
@@ -28,9 +33,14 @@ const updateOneBrand: T.Resolver<GraphQL.MutationUpdateOneBrandArgs, GraphQL.Bra
   return {
     id: result.id,
     url: result.url,
+    v: result.v,
     email: result.email,
-    avgTimeStory: result.avgAllTimeStory || 0,
-    avgAllTimeStory: result.avgAllTimeStory || 0,
+    isPublished: result.isPublished,
+    added: result.added,
+    edited: result.edited,
+    published: result.published,
+    avgTimeStory: result.avgAllTimeStory,
+    avgAllTimeStory: result.avgAllTimeStory,
     created: result.created?.toISOString() || '',
     updated: result.updated?.toISOString() || '',
   };