+++ src/graphql/resolvers/editor/getMany.ts
@@ -14,12 +14,12 @@ const getManyEditor: T.Resolver<void, GraphQL.Editor[]> = async () => {
     return {
       id: result.id,
       name: result.name,
-      edited: result.edited || 0,
-      editedPercent: result.editedPercent || 0,
-      published: result.published || 0,
-      publishedPercent: result.publishedPercent || 0,
-      rejected: result.rejectedPercent || 0,
-      rejectedPercent: result.rejectedPercent || 0,
+      edited: result.edited,
+      editedPercent: result.editedPercent,
+      published: result.published,
+      publishedPercent: result.publishedPercent,
+      rejected: result.rejectedPercent,
+      rejectedPercent: result.rejectedPercent,
       created: result.created?.toISOString() || '',
       updated: result.updated?.toISOString() || '',
     };