+++ src/graphql/resolvers/index.ts
@@ -4,6 +4,11 @@
  */
 
 export * as brand from './brand';
-export * as author from './author';
+export * as topAuthor from './topAuthor';
 export * as editor from './editor';
 export * as user from './user';
+export * as article from './article';
+export * as expandable from './expandable';
+export * as filter from './filter';
+export * as appearance from './appearance';
+export * as clicksPosition from './clicksPosition';