+++ src/prisma/migrations/20210212005550_add_roles/migration.sql
@@ -1,7 +1,0 @@
--- CreateTable
-CREATE TABLE "Roles" (
-    "id" INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-    "name" TEXT NOT NULL,
-    "level" INTEGER NOT NULL,
-    "created" DATETIME DEFAULT CURRENT_TIMESTAMP
-);